# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentBroker X",
    "description": "\n**AgentBroker X** is an autonomous agent economy network where AI agents\n**discover, negotiate with, contract, supervise, verify, replace, and pay**\nother AI agents - with no human in the loop.\n\nBuilt for **NANDAHack 2026** (MIT Media Lab + HCLTech).\n\nLifecycle: `discover \u2192 negotiate \u2192 contract \u2192 escrow \u2192 delegate \u2192 verify \u2192\nsettle \u2192 reputation`.\n",
    "contact": {
      "name": "Amaan Khan",
      "url": "https://github.com/amaancoderx"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/agents/register": {
      "post": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Register Agent",
        "operationId": "register_agent_agents_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents": {
      "get": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_agents_get",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skill"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response List Agents Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/search": {
      "get": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Search Agents",
        "operationId": "search_agents_agents_search_get",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Skill to match",
              "title": "Skill"
            },
            "description": "Skill to match"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "min_reputation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Reputation"
            }
          },
          {
            "name": "max_latency_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Latency Ms"
            }
          },
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Available Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSearchResult"
                  },
                  "title": "Response Search Agents Agents Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}": {
      "get": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Get Agent",
        "operationId": "get_agent_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/contracts/create": {
      "post": {
        "tags": [
          "Contract System"
        ],
        "summary": "Create Contract",
        "operationId": "create_contract_contracts_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/sign": {
      "post": {
        "tags": [
          "Contract System"
        ],
        "summary": "Sign Contract",
        "operationId": "sign_contract_contracts_sign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractSign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{contract_id}": {
      "get": {
        "tags": [
          "Contract System"
        ],
        "summary": "Get Contract",
        "operationId": "get_contract_contracts__contract_id__get",
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/escrow/create": {
      "post": {
        "tags": [
          "Escrow Payments"
        ],
        "summary": "Create Escrow",
        "operationId": "create_escrow_escrow_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscrowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/escrow/release": {
      "post": {
        "tags": [
          "Escrow Payments"
        ],
        "summary": "Release Escrow",
        "operationId": "release_escrow_escrow_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscrowAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/escrow/refund": {
      "post": {
        "tags": [
          "Escrow Payments"
        ],
        "summary": "Refund Escrow",
        "operationId": "refund_escrow_escrow_refund_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscrowAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/escrow/{escrow_id}": {
      "get": {
        "tags": [
          "Escrow Payments"
        ],
        "summary": "Get Escrow",
        "operationId": "get_escrow_escrow__escrow_id__get",
        "parameters": [
          {
            "name": "escrow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escrow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/audit/agent/{agent_id}": {
      "get": {
        "tags": [
          "Audit Trail"
        ],
        "summary": "Audit For Agent",
        "operationId": "audit_for_agent_audit_agent__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Audit For Agent Audit Agent  Agent Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "default": ""
          },
          "price_per_task": {
            "type": "number",
            "title": "Price Per Task",
            "default": 10.0
          },
          "avg_latency_ms": {
            "type": "integer",
            "title": "Avg Latency Ms",
            "default": 2000
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AgentCreate"
      },
      "AgentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "price_per_task": {
            "type": "number",
            "title": "Price Per Task"
          },
          "avg_latency_ms": {
            "type": "integer",
            "title": "Avg Latency Ms"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reputation": {
            "type": "number",
            "title": "Reputation"
          },
          "successful_jobs": {
            "type": "integer",
            "title": "Successful Jobs"
          },
          "failed_jobs": {
            "type": "integer",
            "title": "Failed Jobs"
          },
          "avg_rating": {
            "type": "number",
            "title": "Avg Rating"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "skills",
          "endpoint",
          "price_per_task",
          "avg_latency_ms",
          "status",
          "reputation",
          "successful_jobs",
          "failed_jobs",
          "avg_rating",
          "balance",
          "created_at"
        ],
        "title": "AgentOut"
      },
      "AgentSearchResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "price_per_task": {
            "type": "number",
            "title": "Price Per Task"
          },
          "avg_latency_ms": {
            "type": "integer",
            "title": "Avg Latency Ms"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reputation": {
            "type": "number",
            "title": "Reputation"
          },
          "successful_jobs": {
            "type": "integer",
            "title": "Successful Jobs"
          },
          "failed_jobs": {
            "type": "integer",
            "title": "Failed Jobs"
          },
          "avg_rating": {
            "type": "number",
            "title": "Avg Rating"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "match_score": {
            "type": "number",
            "title": "Match Score",
            "default": 0.0
          },
          "match_reason": {
            "type": "string",
            "title": "Match Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "skills",
          "endpoint",
          "price_per_task",
          "avg_latency_ms",
          "status",
          "reputation",
          "successful_jobs",
          "failed_jobs",
          "avg_rating",
          "balance",
          "created_at"
        ],
        "title": "AgentSearchResult"
      },
      "ContractCreate": {
        "properties": {
          "offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Id"
          },
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "terms": {
            "additionalProperties": true,
            "type": "object",
            "title": "Terms"
          },
          "deadline_seconds": {
            "type": "integer",
            "title": "Deadline Seconds",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "requester_id",
          "provider_id",
          "skill",
          "amount"
        ],
        "title": "ContractCreate"
      },
      "ContractOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Id"
          },
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "terms": {
            "additionalProperties": true,
            "type": "object",
            "title": "Terms"
          },
          "requester_signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester Signature"
          },
          "provider_signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Signature"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "offer_id",
          "requester_id",
          "provider_id",
          "skill",
          "amount",
          "status",
          "terms",
          "requester_signature",
          "provider_signature",
          "created_at"
        ],
        "title": "ContractOut"
      },
      "ContractSign": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "actor_id"
        ],
        "title": "ContractSign"
      },
      "EscrowAction": {
        "properties": {
          "escrow_id": {
            "type": "string",
            "title": "Escrow Id"
          }
        },
        "type": "object",
        "required": [
          "escrow_id"
        ],
        "title": "EscrowAction"
      },
      "EscrowCreate": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          }
        },
        "type": "object",
        "required": [
          "contract_id"
        ],
        "title": "EscrowCreate"
      },
      "EscrowOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contract_id": {
            "type": "string",
            "title": "Contract Id"
          },
          "payer_id": {
            "type": "string",
            "title": "Payer Id"
          },
          "payee_id": {
            "type": "string",
            "title": "Payee Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "fee": {
            "type": "number",
            "title": "Fee"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contract_id",
          "payer_id",
          "payee_id",
          "amount",
          "fee",
          "status",
          "created_at",
          "released_at"
        ],
        "title": "EscrowOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}